implement the acl to marketing.

+++ packages/Webkul/Admin/src/DataGrids/Settings/Marketing/CampaignDatagrid.php
@@ -65,22 +65,25 @@ class CampaignDatagrid extends DataGrid
      */
     public function prepareActions()
     {
-        // TODO: Implement the bouncer for this actions.
-        $this->addAction([
-            'index'  => 'edit',
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.settings.marketing.campaigns.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => fn ($row) => route('admin.settings.marketing.events.edit', $row->id),
-        ]);
+        if (bouncer()->hasPermission('settings.automation.campaigns.edit')) {
+            $this->addAction([
+                'index'  => 'edit',
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.settings.marketing.campaigns.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => fn ($row) => route('admin.settings.marketing.events.edit', $row->id),
+            ]);
+        }
 
-        $this->addAction([
-            'index'          => 'delete',
-            'icon'           => 'icon-delete',
-            'title'          => trans('admin::app.settings.marketing.campaigns.index.datagrid.delete'),
-            'method'         => 'DELETE',
-            'url'            => fn ($row) => route('admin.settings.marketing.events.delete', $row->id),
-        ]);
+        if (bouncer()->hasPermission('settings.automation.campaigns.delete')) {
+            $this->addAction([
+                'index'          => 'delete',
+                'icon'           => 'icon-delete',
+                'title'          => trans('admin::app.settings.marketing.campaigns.index.datagrid.delete'),
+                'method'         => 'DELETE',
+                'url'            => fn ($row) => route('admin.settings.marketing.events.delete', $row->id),
+            ]);
+        }
     }
 
     /**
@@ -88,11 +91,13 @@ class CampaignDatagrid extends DataGrid
      */
     public function prepareMassActions(): void
     {
-        $this->addMassAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.settings.marketing.campaigns.index.datagrid.delete'),
-            'method' => 'POST',
-            'url'    => route('admin.settings.marketing.events.mass_delete'),
-        ]);
+        if (bouncer()->hasPermission('settings.automation.campaigns.mass_delete')) {
+            $this->addMassAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.settings.marketing.campaigns.index.datagrid.delete'),
+                'method' => 'POST',
+                'url'    => route('admin.settings.marketing.events.mass_delete'),
+            ]);
+        }
     }
 }
